Move testhelpers into shared/osarch for now
This avoids us an extra build dependency on stretchr.
